pyproject.toml
[project]
name = "alphavantage-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "firecrawl-py>=2.12.0",
    "google-genai>=1.23.0",
    "langchain[google-genai]>=0.3.26",
    "langchain-google-genai>=2.1.6",
    "langchain-mcp-adapters>=0.1.8",
    "langgraph>=0.5.0",
    "mcp>=1.10.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
]
